Separate entrypoints for static & editable diagrams
Static diagram doesn't require most components used
by the editable diagram. By separating the entry point
we can easily have a less sized build for apps which
only need static diagram to be rendered. For example,
the playground widget on ballerina.io.
